Move state to app and adjust condtional rendering
